Install firebase with npm
Installation through bower does not let us require('firebase') on node
to run tests.
Could not find documentation on how to do that, instead found -
https://gofore.com/ohjelmistokehitys/stop-using-bower/

So a clear split for now:
- purescript, pulp, bower are installed through npm

- purescript-* are installed through bower.

Some work (browserify or webpack) still needs to be done to use firebase in
the browser with this setup.