Once error callback finally throws exception
Seems to have been missing () function call of the callback (there was a
hack for that on the success side, the types did not line up well
enough).

Perhaps doing a bit more on the JS side to keep the types nice in PS.
Process callbacks right the first time (maybe it is because of runFnX,
instead of writing by hand. It was not obvious to spot that the error
callback was actually not called).