Propagate errors for on() in Aff
Before this, it would just hang indefinetely. Effects of callbacks need
to be called (the Aff documentation says as much, but an innocent extra
pair of () didn't jump out at me).

For now resolved it on the purescript side, so it is a bit more clear
this needs to be done, even though the types are not pretty and we incur
an additional if for each callback.