Simplified error handling for on and once
Error callbacks are not as optional as the documentation suggests: if
you don't provide one, firebase will throw an error, and it is likely to
kill your programme (in our case, no tests run. I don't feel like adding
Exception as an effect, we have error callbacks and Eff to deal with
that). If you don't want to thin too much about errors, use the Aff
implementations.