Remove all json parsing /  foreign conversions.
Data.IsForeign got removed upstream, maybe it is a signal there are too
many responsibilities in this library.

The authentication code should probably go to another repository, as it
is likely to incur a dependency on a specific form of generics, where
the others are not.

FYI user credentials are also affected by loading / not yet loaded,
which is not reflected in the code yet.